extract common overlay util
