+++ src/overlay/index.html
@@ -16,6 +16,6 @@
         </style>
         <script src="index.js"></script>
     </head>
-    <body class="click-through etmf-void">
+    <body class="click-through etmf-void" style="background-color: rgba(0, 0, 0, 0.05);">
     </body>
 </html>
